perf: generate history on build (1)
Move `getHistory()` from utils/utils.ts to scripts/build.ts as `generateHistoryJson()`
